+++ cypress/e2e/WorkingLists/EventWorkingLists/EventWorkingListsUser/EventWorkingListsUser.js
@@ -3,8 +3,16 @@ import { v4 as uuid } from 'uuid';
 import '../sharedSteps';
 import { combineDataAndYear, getCurrentYear } from '../../../../support/date';
 
-Given('you open the main page with Ngelehun and malaria case context', () => {
-    cy.visit('#/?programId=VBqh0ynB2wv&orgUnitId=DiszpKrYNg8');
+const CONTEXT_QUERIES = {
+    'malaria case context': 'programId=VBqh0ynB2wv&orgUnitId=DiszpKrYNg8',
+    'Inpatient morbidity and mortality context': 'programId=eBAyeGv0exc&orgUnitId=DiszpKrYNg8',
+    'Contraceptives Voucher Program': 'orgUnitId=DiszpKrYNg8&programId=kla3mAPgvCH',
+    'event program text filter context': 'programId=MoUd5BTQ3lY&orgUnitId=DiszpKrYNg8',
+};
+
+Given(/^you open the main page with Ngelehun and (.+)$/, (contextOrPath) => {
+    const query = CONTEXT_QUERIES[contextOrPath] ?? contextOrPath;
+    cy.visit(`#/?${query}`);
 });
 
 Then('the default working list should be displayed', () => {
@@ -138,6 +146,97 @@ Then('the list should display events where age is between 10 and 20', () => {
                     .contains(rows[index - 1].split(' ')[1])
                     .should('exist');
             }
+        });
+});
+
+When('you set the Household location filter to Is empty', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .within(() => {
+            cy.contains('More filters')
+                .click();
+        });
+
+    cy.get('[data-test="more-filters-menu"]')
+        .within(() => cy.contains('Household location').click());
+
+    cy.get('[data-test="list-view-filter-contents"]')
+        .contains('Is empty')
+        .click();
+});
+
+Then('the Household location filter button should show that the filter is in effect', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Household location: Is empty')
+        .should('exist');
+});
+
+Then('the age filter button should show that the filter is in effect', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains(/Age \(years\): \d+ to \d+/)
+        .should('exist');
+});
+
+When(/^you set the text filter "([^"]+)" to "([^"]+)"$/, (filterName, value) => {
+    cy.get('[data-test="event-working-lists"]')
+        .within(() => {
+            cy.contains('More filters').click();
+        });
+    cy.get('[data-test="more-filters-menu"]')
+        .within(() => cy.contains(filterName).click());
+    cy.get('[data-test="list-view-filter-contents"]')
+        .find('input[type="text"]')
+        .clear()
+        .type(value)
+        .blur();
+    cy.get('[data-test="list-view-filter-apply-button"]').click();
+});
+
+Then(/^the text filter "([^"]+)" should be in effect and show "([^"]+)" when opened$/, (filterName, value) => {
+    cy.get('[data-test="event-working-lists"]').contains(`${filterName}: ${value}`).should('exist');
+    cy.get('[data-test="event-working-lists"]').contains(filterName).click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.get('input[type="text"]').should('have.value', value);
+    });
+    cy.get('body').click(0, 0);
+});
+
+Then('the Household location filter should show Is empty checked', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Household location')
+        .click();
+    cy.get('[data-test="list-view-filter-contents"]')
+        .within(() => {
+            cy.contains('Is empty')
+                .closest('label')
+                .find('input[type="checkbox"]')
+                .should('be.checked');
+        });
+    cy.get('body').click(0, 0);
+});
+
+Then('the list should display one record', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .within(() => {
+            cy.get('[data-test="working-list-table-loading"]').should('not.exist');
+        });
+    cy.get('[data-test="event-working-lists"]')
+        .find('tr')
+        .should('have.length', 2);
+});
+
+Then('the list should display one record with empty Household location', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .within(() => {
+            cy.get('[data-test="working-list-table-loading"]').should('not.exist');
+        });
+    cy.get('[data-test="event-working-lists"]')
+        .find('tr')
+        .should('have.length', 2);
+    cy.get('[data-test="event-working-lists"]')
+        .find('tr')
+        .eq(1)
+        .within(() => {
+            cy.get('td').contains('Lat:').should('not.exist');
         });
 });
 
@@ -334,25 +433,6 @@ When('you create a copy of the working list', () => {
         .click();
 });
 
-When('you change the sharing settings', () => {
-    cy.get('[data-test="list-view-menu-button"]')
-        .click();
-
-    cy.contains('Share view')
-        .click();
-    cy.get('[placeholder="Search"]')
-        .type('Boateng');
-
-    cy.contains('Kevin Boateng').click();
-    cy.contains('Choose a level').click();
-    cy.get('[data-test="dhis2-uicore-select-menu-menuwrapper"]')
-        .contains('View and edit').click({ force: true });
-
-    cy.get('[data-test="dhis2-uicore-button"]').contains('Give access').click({ force: true });
-    cy.get('[data-test="dhis2-uicore-button"]').contains('Close').click({ force: true });
-});
-
-
 When('you update the working list', () => {
     cy.get('[data-test="dhis2-uicore-tableheadercellaction"]')
         .eq(0)
@@ -390,10 +470,6 @@ Then('your newly defined sharing settings should still be present', () => {
     cy.contains('Confirm')
         .click();
 });
-Given('you open the main page with Ngelehun and Inpatient morbidity and mortality context', () => {
-    cy.visit('#/?programId=eBAyeGv0exc&orgUnitId=DiszpKrYNg8');
-});
-
 When('you set the date of admission filter', () => {
     cy.get('[data-test="event-working-lists"]')
         .within(() => {
@@ -402,7 +478,7 @@ When('you set the date of admission filter', () => {
         });
 
     cy.get('[data-test="more-filters-menu"]')
-        .within(() => cy.contains('Date of admission').click());
+        .within(() => cy.contains(/Date of admission|Admission Date/).click());
 
     cy.get('[data-test="list-view-filter-contents"]')
         .within(() => {
@@ -418,6 +494,44 @@ When('you set the date of admission filter', () => {
                 });
 
             cy.contains('Update')
+                .click();
+        });
+});
+
+When('you set the report date filter', () => {
+    const year = getCurrentYear() - 1;
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Report date')
+        .click();
+
+    cy.get('[data-test="list-view-filter-contents"]')
+        .within(() => {
+            cy.contains('Absolute range')
+                .click();
+            cy.get('input[type="text"]')
+                .then(($elements) => {
+                    cy.wrap($elements[0])
+                        .type(`${year}-01-01`).blur();
+
+                    cy.wrap($elements[1])
+                        .type(`${year}-12-31`).blur();
+                });
+
+            cy.contains('Update')
+                .click();
+        });
+});
+
+When('you set the Organisation unit filter to Ngelehun', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Organisation unit')
+        .click();
+
+    cy.get('[data-test="list-view-filter-contents"]')
+        .within(() => {
+            cy.get('input[placeholder="Search"]')
+                .type('Ngelehun');
+            cy.contains('Ngelehun', { timeout: 10000 })
                 .click();
         });
 });
@@ -445,14 +559,435 @@ When(/^you save the view as (.*)$/, (name) => {
     cy.wait('@newEventFilter', { timeout: 30000 }).as('newEventResult');
 });
 
-When('you refresh the page', () => {
-    cy.reload();
+When('you set the Pregnant filter to Yes', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .within(() => cy.contains('More filters').click());
+    cy.get('[data-test="more-filters-menu"]')
+        .within(() => cy.contains('Pregnant').click());
+    cy.get('[data-test="list-view-filter-contents"]')
+        .contains('Yes')
+        .click();
+    cy.get('[data-test="list-view-filter-apply-button"]').click();
 });
 
-When('you open the dateFilterWorkingList', () => {
-    cy.get('[data-test="workinglists-template-selector-chips-container"]')
-        .contains('dateFilterWorkingList')
+When('you set the Age (years) filter to 0-120', () => {
+    cy.get('[data-test="event-working-lists"]').contains('Age (years)').click();
+    cy.get('[data-test="list-view-filter-contents"]')
+        .find('input[placeholder="Min"]')
+        .type('0')
+        .blur();
+    cy.get('[data-test="list-view-filter-contents"]')
+        .find('input[placeholder="Max"]')
+        .type('120')
+        .blur();
+    cy.get('[data-test="list-view-filter-apply-button"]').click();
+});
+
+When('you set the Height in cm filter to 100-200', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .within(() => cy.contains('More filters').click());
+    cy.get('[data-test="more-filters-menu"]')
+        .within(() => cy.contains('Height in cm').click());
+    cy.get('[data-test="list-view-filter-contents"]')
+        .find('input[placeholder="Min"]')
+        .type('100')
+        .blur();
+    cy.get('[data-test="list-view-filter-contents"]')
+        .find('input[placeholder="Max"]')
+        .type('200')
+        .blur();
+    cy.get('[data-test="list-view-filter-apply-button"]').click();
+});
+
+When('you set the Weight in kg filter to 1-200', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .within(() => cy.contains('More filters').click());
+    cy.get('[data-test="more-filters-menu"]')
+        .within(() => cy.contains('Weight in kg').click());
+    cy.get('[data-test="list-view-filter-contents"]')
+        .find('input[placeholder="Min"]')
+        .type('1')
+        .blur();
+    cy.get('[data-test="list-view-filter-contents"]')
+        .find('input[placeholder="Max"]')
+        .type('200')
+        .blur();
+    cy.get('[data-test="list-view-filter-apply-button"]').click();
+});
+
+When('you set the Place of Infection filter to Ngelehun', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .within(() => cy.contains('More filters').click());
+    cy.get('[data-test="more-filters-menu"]')
+        .within(() => cy.contains('Place of Infection').click());
+    cy.get('[data-test="list-view-filter-contents"]')
+        .within(() => {
+            cy.get('input[placeholder="Search"]').type('Ngelehun');
+            cy.contains('Ngelehun', { timeout: 10000 }).click();
+        });
+});
+
+When(/^you set the range filter "([^"]+)" to (\d+)-(\d+)$/, (filterName, min, max) => {
+    if (filterName === 'Age (years)') {
+        cy.get('[data-test="event-working-lists"]').contains('Age (years)').click();
+    } else {
+        cy.get('[data-test="event-working-lists"]').within(() => cy.contains('More filters').click());
+        cy.get('[data-test="more-filters-menu"]').within(() => cy.contains(filterName).click());
+    }
+    cy.get('[data-test="list-view-filter-contents"]').find('input[placeholder="Min"]').type(min).blur();
+    cy.get('[data-test="list-view-filter-contents"]').find('input[placeholder="Max"]').type(max).blur();
+    cy.get('[data-test="list-view-filter-apply-button"]').click();
+});
+
+When('you set the boolean filter', () => {
+    cy.get('[data-test="event-working-lists"]').within(() => cy.contains('More filters').click());
+    cy.get('[data-test="more-filters-menu"]').within(() => cy.contains('Pregnant').click());
+    cy.get('[data-test="list-view-filter-contents"]').contains('Yes').click();
+    cy.get('[data-test="list-view-filter-apply-button"]').click();
+});
+
+When('you set the date filter', () => {
+    cy.get('[data-test="event-working-lists"]').within(() => cy.contains('More filters').click());
+    cy.get('[data-test="more-filters-menu"]').within(() => cy.contains(/Date of admission|Admission Date/).click());
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.contains('Absolute range').click();
+        cy.get('input[type="text"]').then(($inputs) => {
+            cy.wrap($inputs[0]).type('2018-01-01').blur();
+            cy.wrap($inputs[1]).type('2018-12-31').blur();
+        });
+        cy.contains('Update').click();
+    });
+});
+
+When('you set the organisation unit filter', () => {
+    cy.get('[data-test="event-working-lists"]').within(() => cy.contains('More filters').click());
+    cy.get('[data-test="more-filters-menu"]').within(() => cy.contains('Place of Infection').click());
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.get('input[placeholder="Search"]').type('Ngelehun');
+        cy.get('[data-test="dhis2-uicore-circularloader"]').should('not.exist');
+        cy.contains('Ngelehun').click();
+    });
+});
+
+When(/^you set the empty-only filter "([^"]+)" to (Is empty|Is not empty)$/, (filterName, value) => {
+    cy.get('[data-test="event-working-lists"]').within(() => cy.contains('More filters').click());
+    cy.get('[data-test="more-filters-menu"]').within(() => cy.contains(filterName).click());
+    cy.get('[data-test="list-view-filter-contents"]').contains(value).click();
+    cy.get('[data-test="list-view-filter-apply-button"]').click();
+});
+
+Then('all set filters should show in effect', () => {
+    cy.get('[data-test="event-working-lists"]').should('contain', 'Pregnant').and('contain', 'Yes');
+    cy.get('[data-test="event-working-lists"]').contains('Age (years): 0 to 120').should('exist');
+    cy.get('[data-test="event-working-lists"]').contains('Height in cm: 100 to 200').should('exist');
+    cy.get('[data-test="event-working-lists"]').contains('Weight in kg: 1 to 200').should('exist');
+    cy.get('[data-test="event-working-lists"]').should(($el) => {
+        expect($el.text()).to.include('2018');
+        expect($el.text()).to.match(/Admission Date|Date of admission/);
+    });
+    cy.get('[data-test="event-working-lists"]').should('contain', 'Place of Infection').and('contain', 'Ngelehu');
+    cy.get('[data-test="event-working-lists"]').contains('Household location: Is empty').should('exist');
+    cy.get('[data-test="event-working-lists"]').contains('Documentation: Is empty').should('exist');
+});
+
+Then('the boolean filter should be in effect and show the correct value when opened', () => {
+    cy.get('[data-test="event-working-lists"]').should('contain', 'Pregnant').and('contain', 'Yes');
+    cy.get('[data-test="event-working-lists"]').contains('Pregnant').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.contains('Yes').closest('label').find('input').should('be.checked');
+    });
+    cy.get('body').click(0, 0);
+});
+
+Then(/^the range filter "([^"]+)" should be in effect and show (\d+) to (\d+) when opened$/, (filterName, min, max) => {
+    cy.get('[data-test="event-working-lists"]').contains(`${filterName}: ${min} to ${max}`).should('exist');
+    cy.get('[data-test="event-working-lists"]').contains(filterName).click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.get('input[placeholder="Min"]').should('have.attr', 'value', min);
+        cy.get('input[placeholder="Max"]').should('have.attr', 'value', max);
+    });
+    cy.get('body').click(0, 0);
+});
+
+Then('the date filter should be in effect and show the correct value when opened', () => {
+    cy.get('[data-test="event-working-lists"]').should(($el) => {
+        expect($el.text()).to.include('2018');
+        expect($el.text()).to.match(/Admission Date|Date of admission/);
+    });
+    cy.get('[data-test="event-working-lists"]').contains(/Admission Date|Date of admission/).click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.contains('Absolute range').click();
+        cy.get('input[type="text"]').then(($inputs) => {
+            cy.wrap($inputs[0]).should('have.attr', 'value', '2018-01-01');
+            cy.wrap($inputs[1]).should('have.attr', 'value', '2018-12-31');
+        });
+    });
+    cy.get('body').click(0, 0);
+});
+
+Then('the organisation unit filter should be in effect and show the correct value when opened', () => {
+    cy.get('[data-test="event-working-lists"]').should('contain', 'Place of Infection').and('contain', 'Ngelehu');
+    cy.get('[data-test="event-working-lists"]').contains('Place of Infection').click();
+    cy.get('[data-test="list-view-filter-contents"]').should('contain', 'Ngelehun');
+    cy.get('body').click(0, 0);
+});
+
+Then(/^the empty-only filter "([^"]+)" should be in effect and show (Is empty|Is not empty) when opened$/, (filterName, value) => {
+    cy.get('[data-test="event-working-lists"]').contains(`${filterName}: ${value}`).should('exist');
+    cy.get('[data-test="event-working-lists"]').contains(filterName).click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.contains(value).closest('label').find('input[type="checkbox"]').should('be.checked');
+    });
+    cy.get('body').click(0, 0);
+});
+
+Then('the boolean and range filters should show correct values when opened', () => {
+    cy.get('[data-test="event-working-lists"]').contains('Pregnant').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.contains('Yes').closest('label').find('input').should('be.checked');
+    });
+    cy.get('body').click(0, 0);
+
+    cy.get('[data-test="event-working-lists"]').contains('Age (years)').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.get('input[placeholder="Min"]').should('have.attr', 'value', '0');
+        cy.get('input[placeholder="Max"]').should('have.attr', 'value', '120');
+    });
+    cy.get('body').click(0, 0);
+
+    cy.get('[data-test="event-working-lists"]').contains('Height in cm').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.get('input[placeholder="Min"]').should('have.attr', 'value', '100');
+        cy.get('input[placeholder="Max"]').should('have.attr', 'value', '200');
+    });
+    cy.get('body').click(0, 0);
+
+    cy.get('[data-test="event-working-lists"]').contains('Weight in kg').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.get('input[placeholder="Min"]').should('have.attr', 'value', '1');
+        cy.get('input[placeholder="Max"]').should('have.attr', 'value', '200');
+    });
+    cy.get('body').click(0, 0);
+});
+
+Then('the date and coordinate filters should show correct values when opened', () => {
+    cy.get('[data-test="event-working-lists"]').contains(/Admission Date|Date of admission/).click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.contains('Absolute range').click();
+        cy.get('input[type="text"]').then(($inputs) => {
+            cy.wrap($inputs[0]).should('have.attr', 'value', '2018-01-01');
+            cy.wrap($inputs[1]).should('have.attr', 'value', '2018-12-31');
+        });
+    });
+    cy.get('body').click(0, 0);
+
+    cy.get('[data-test="event-working-lists"]').contains('Household location').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.contains('Is empty').closest('label').find('input[type="checkbox"]').should('be.checked');
+    });
+    cy.get('body').click(0, 0);
+});
+
+Then('each filter should show correct value when opened', () => {
+    cy.get('[data-test="event-working-lists"]').contains('Pregnant').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.contains('Yes').closest('label').find('input').should('be.checked');
+    });
+    cy.get('body').click(0, 0);
+
+    cy.get('[data-test="event-working-lists"]').contains('Age (years)').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.get('input[placeholder="Min"]').should('have.attr', 'value', '0');
+        cy.get('input[placeholder="Max"]').should('have.attr', 'value', '120');
+    });
+    cy.get('body').click(0, 0);
+
+    cy.get('[data-test="event-working-lists"]').contains('Height in cm').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.get('input[placeholder="Min"]').should('have.attr', 'value', '100');
+        cy.get('input[placeholder="Max"]').should('have.attr', 'value', '200');
+    });
+    cy.get('body').click(0, 0);
+
+    cy.get('[data-test="event-working-lists"]').contains('Weight in kg').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.get('input[placeholder="Min"]').should('have.attr', 'value', '1');
+        cy.get('input[placeholder="Max"]').should('have.attr', 'value', '200');
+    });
+    cy.get('body').click(0, 0);
+
+    cy.get('[data-test="event-working-lists"]').contains(/Admission Date|Date of admission/).click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.contains('Absolute range').click();
+        cy.get('input[type="text"]').then(($inputs) => {
+            cy.wrap($inputs[0]).should('have.attr', 'value', '2018-01-01');
+            cy.wrap($inputs[1]).should('have.attr', 'value', '2018-12-31');
+        });
+    });
+    cy.get('body').click(0, 0);
+
+    cy.get('[data-test="event-working-lists"]').contains('Household location').click();
+    cy.get('[data-test="list-view-filter-contents"]').within(() => {
+        cy.contains('Is empty').closest('label').find('input[type="checkbox"]').should('be.checked');
+    });
+    cy.get('body').click(0, 0);
+});
+
+Then('the Pregnant filter should show Yes in effect', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .should('contain', 'Pregnant')
+        .and('contain', 'Yes');
+});
+
+Then('the Age (years) filter should show 0 to 120 in effect', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Age (years): 0 to 120')
+        .should('exist');
+});
+
+Then('the Height in cm filter should show 100 to 200 in effect', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Height in cm: 100 to 200')
+        .should('exist');
+});
+
+Then('the Weight in kg filter should show 1 to 200 in effect', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Weight in kg: 1 to 200')
+        .should('exist');
+});
+
+Then('the Admission Date filter should show date range in effect', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .should(($el) => {
+            const text = $el.text();
+            expect(text).to.include('2018');
+            expect(text).to.match(/Admission Date|Date of admission/);
+        });
+});
+
+Then('the Place of Infection filter should show Ngelehun in effect', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .should('contain', 'Place of Infection')
+        .and('contain', 'Ngelehun');
+});
+
+Then('the Pregnant filter should show Yes radio checked when opened', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Pregnant')
         .click();
+    cy.get('[data-test="list-view-filter-contents"]')
+        .within(() => {
+            cy.contains('Yes')
+                .closest('label')
+                .find('input')
+                .should('be.checked');
+        });
+    cy.get('body').click(0, 0);
+});
+
+Then('the Age (years) filter should show 0 and 120 in range fields when opened', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Age (years)')
+        .click();
+    cy.get('[data-test="list-view-filter-contents"]')
+        .within(() => {
+            cy.get('input[placeholder="Min"]').should('have.attr', 'value', '0');
+            cy.get('input[placeholder="Max"]').should('have.attr', 'value', '120');
+        });
+    cy.get('body').click(0, 0);
+});
+
+Then('the Height in cm filter should show 100 and 200 in range fields when opened', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Height in cm')
+        .click();
+    cy.get('[data-test="list-view-filter-contents"]')
+        .within(() => {
+            cy.get('input[placeholder="Min"]').should('have.attr', 'value', '100');
+            cy.get('input[placeholder="Max"]').should('have.attr', 'value', '200');
+        });
+    cy.get('body').click(0, 0);
+});
+
+Then('the Weight in kg filter should show 1 and 200 in range fields when opened', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Weight in kg')
+        .click();
+    cy.get('[data-test="list-view-filter-contents"]')
+        .within(() => {
+            cy.get('input[placeholder="Min"]').should('have.attr', 'value', '1');
+            cy.get('input[placeholder="Max"]').should('have.attr', 'value', '200');
+        });
+    cy.get('body').click(0, 0);
+});
+
+Then('the Admission Date filter should show date range when opened', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains(/Admission Date|Date of admission/)
+        .click();
+    cy.get('[data-test="list-view-filter-contents"]')
+        .within(() => {
+            cy.contains('Absolute range').click();
+            cy.get('input[type="text"]')
+                .then(($inputs) => {
+                    cy.wrap($inputs[0]).should('have.attr', 'value', '2018-01-01');
+                    cy.wrap($inputs[1]).should('have.attr', 'value', '2018-12-31');
+                });
+        });
+    cy.get('body').click(0, 0);
+});
+
+Then('the list should display one record with report date matching filter', () => {
+    const year = getCurrentYear() - 1;
+    cy.get('[data-test="event-working-lists"]')
+        .within(() => {
+            cy.get('[data-test="working-list-table-loading"]').should('not.exist');
+        });
+    cy.get('[data-test="event-working-lists"]')
+        .find('tr')
+        .should('have.length', 2);
+    cy.get('[data-test="event-working-lists"]')
+        .find('tr')
+        .eq(1)
+        .find('td')
+        .eq(1)
+        .invoke('text')
+        .then((text) => {
+            expect(text).to.include(String(year));
+        });
+});
+
+Then('the report date filter should be in effect', () => {
+    const year = getCurrentYear() - 1;
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Report date')
+        .click();
+
+    cy.get('[data-test="list-view-filter-contents"]')
+        .within(() => {
+            cy.contains('Absolute range')
+                .click();
+            cy.get('input[type="text"]')
+                .then(($elements) => {
+                    cy.wrap($elements[0])
+                        .should('have.attr', 'value', `${year}-01-01`);
+
+                    cy.wrap($elements[1])
+                        .should('have.attr', 'value', `${year}-12-31`);
+                });
+        });
+
+    cy.get('body').click(0, 0);
+});
+
+Then('the Organisation unit filter should be in effect', () => {
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Organisation unit')
+        .should('exist');
+    cy.get('[data-test="event-working-lists"]')
+        .contains('Ngelehun')
+        .should('exist');
 });
 
 Then('the admission filter should be in effect', () => {
@@ -486,8 +1021,16 @@ Then('the admission filter should be in effect', () => {
         });
 });
 
-Given('you open the main page with Ngelehun and Contraceptives Voucher Program', () => {
-    cy.visit('#/?orgUnitId=DiszpKrYNg8&programId=kla3mAPgvCH');
+Then('the saved working list view is cleaned up', () => {
+    cy.get('@newEventResult')
+        .then((result) => {
+            expect(result.response.statusCode).to.equal(201);
+            const id = result.response.body.response.uid;
+            cy.buildApiUrl('eventFilters', id)
+                .then((eventFiltersUrl) => {
+                    cy.request('DELETE', eventFiltersUrl);
+                });
+        });
 });
 
 Then(/^you are told to select (.*)$/, (text) => {