+++ src/core_modules/capture-core/components/D2Form/field/configs/textField/getTextFieldConfig.ts
@@ -18,7 +18,6 @@ export const getTextFieldConfig = (
         formHorizontal: options.formHorizontal,
         fieldLabelMediaBasedClass: options.fieldLabelMediaBasedClass,
         multiLine: extraProps && extraProps.multiLine,
-        showHelpText: options.showHelpText,
         ...(isLtrInputType(metaData.type) ? { dir: 'ltr' as const } : {}),
     }, options, metaData);