+++ src/core_modules/capture-core/components/FiltersForTypes/OrgUnit/OrgUnitFilter.component.tsx
@@ -19,7 +19,9 @@ export class OrgUnitFilter extends Component<OrgUnitFilterProps> implements Upda
 
     handleOrgUnitChange = (value: OrgUnitValue | null) => {
         this.props.onCommitValue(value);
-        if (value) {
+        if (value === null && this.props.onClearValue) {
+            this.props.onClearValue();
+        } else {
             this.props.onUpdate(value);
         }
     };