+++ src/core_modules/capture-core/components/ListView/Filters/FilterButton/buttonTextBuilder/converters/numericConverter.ts
@@ -2,22 +2,28 @@ import i18n from '@dhis2/d2-i18n';
 import type { NumericFilterData } from '../../../../../FiltersForTypes';
 
 export function convertNumeric(filter: NumericFilterData): string {
-    if (filter.value) {
-        return filter.value;
+    let appliedText = '';
+    const geHasValue = !!filter.ge || filter.ge === 0;
+    const leHasValue = !!filter.le || filter.le === 0;
+
+    if (geHasValue && leHasValue) {
+        if (filter.ge === filter.le) {
+            appliedText = String(filter.ge);
+        } else {
+            appliedText = i18n.t('{{from}} to {{to}}', {
+                from: filter.ge,
+                to: filter.le,
+            });
+        }
+    } else if (geHasValue) {
+        appliedText = i18n.t('greater than or equal to {{value}}', {
+            value: filter.ge,
+        });
+    } else {
+        appliedText = i18n.t('less than or equal to {{value}}', {
+            value: filter.le,
+        });
     }
 
-    const { ge, le } = filter;
-    const hasGe = typeof ge === 'number';
-    const hasLe = typeof le === 'number';
-
-    if (hasGe && hasLe) {
-        return ge === le ? String(ge) : `${ge} ${i18n.t('to')} ${le}`;
-    }
-    if (hasGe) {
-        return `${i18n.t('greater than or equal to')} ${ge}`;
-    }
-    if (hasLe) {
-        return `${i18n.t('less than or equal to')} ${le}`;
-    }
-    return '';
+    return appliedText;
 }