+++ src/core_modules/capture-core/components/ListView/Filters/filters.const.ts
@@ -27,3 +27,10 @@ export const filterTypesObject = {
     URL: dataElementTypes.URL,
     USERNAME: dataElementTypes.USERNAME,
 } as const;
+
+export const EMPTY_ONLY_FILTER_TYPES: Set<string> = new Set([
+    filterTypesObject.COORDINATE,
+    filterTypesObject.FILE_RESOURCE,
+    filterTypesObject.IMAGE,
+    filterTypesObject.URL,
+]);