+++ src/core_modules/capture-core/components/WorkingLists/EventWorkingLists/helpers/eventFilters/clientConfigToApiEventFilterQueryCriteriaConverter/convertToEventFilterEventQueryCriteria.ts
@@ -54,8 +54,8 @@ const getOrgUnitFilter = (filter: OrgUnitFilterData): ApiDataFilterOrgUnit => ({
 });
 
 const getNumericFilter = (filter: NumericFilterData): ApiDataFilterNumeric => ({
-    ge: filter.ge ? filter.ge.toString() : undefined,
-    le: filter.le ? filter.le.toString() : undefined,
+    ge: filter.ge?.toString() ?? undefined,
+    le: filter.le?.toString() ?? undefined,
 });
 
 const getTimeFilter = (filter: TimeFilterData): ApiDataFilterNumeric => ({
@@ -216,6 +216,7 @@ const structureFilters = (apiFilters: Array<any>, columns: ColumnsForConverter) 
             if (mainFilter) {
                 return { ...acc, ...mainFilter };
             }
+            return acc;
         }
 
         acc.dataFilters.push(filter);