+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/common/factory/searchGroup/SearchGroupFactory.ts
@@ -155,7 +155,7 @@ export class SearchGroupFactory {
             (acc, attr) => {
                 const valueType = attr.trackedEntityAttribute?.valueType;
                 const isUnsupported = valueType && UNSUPPORTED_SEARCH_ATTRIBUTE_TYPES.has(valueType as any);
-                if (isUnsupported) {
+                if (key === 'main' && isUnsupported) {
                     acc.unsupportedAttributes.push(attr);
                 } else {
                     acc.supportedAttributes.push(attr);
@@ -172,6 +172,8 @@ export class SearchGroupFactory {
         if (key === 'main') {
             searchGroup.minAttributesRequiredToSearch = minAttributesRequiredToSearch;
             searchGroup.id = 'main';
+        } else if (key === 'nonSearchable') {
+            searchGroup.id = 'nonSearchable';
         } else {
             searchGroup.unique = true;
             searchGroup.id = 'unique';
@@ -204,12 +206,16 @@ export class SearchGroupFactory {
                 trackedEntityAttribute: this.getTrackedEntityAttribute(attribute),
             }))
             .filter(attribute =>
-                attribute.trackedEntityAttribute && (attribute.searchable || attribute.trackedEntityAttribute.unique))
+                attribute.trackedEntityAttribute)
             .reduce((accGroups: any, attribute) => {
                 if (attribute.trackedEntityAttribute!.unique) {
                     accGroups[attribute.trackedEntityAttribute!.id] = [attribute];
-                } else {
+                } else if (attribute.searchable) {
                     accGroups.main = accGroups.main ? [...accGroups.main, attribute] : [attribute];
+                } else {
+                    accGroups.nonSearchable = accGroups.nonSearchable
+                        ? [...accGroups.nonSearchable, attribute]
+                        : [attribute];
                 }
                 return accGroups;
             }, {});